refactor(client): move SignUp component to `pages`
